requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.9
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
gitpython==3.1.45
greenlet==3.2.4
griffe==1.15.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.8
llama-index-cli==0.5.3
llama-index-core==0.14.8
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-ollama==0.9.0
llama-index-llms-openai==0.6.9
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.11.1
llama-parse==0.6.54
lxml==6.0.2
markupsafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
mypy-extensions==1.1.0
narwhals==2.11.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
ollama==0.6.1
openai==2.8.0
packaging==25.0
pandas==2.2.3
pillow==12.0.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.33.1
pyarrow==21.0.0
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.3.4
pydeck==0.9.1
pypdf==6.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.44
streamlit==1.51.0
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
wrapt==1.17.3
yarl==1.22.0
